fix: Jetzt wird wirklich der hash genommen und nicht die byte size.

+++ internal/services/media.go
@@ -2,6 +2,6 @@ package services
 
 import "github.com/bestreads/Backend/internal/database"
 
-func SaveFile(data []byte) (int, error) {
+func SaveFile(data []byte) (uint64, error) {
 	return database.FileStoreRaw(data)
 }
